# docs/swagger.yaml
definitions:
...
  requestmodels.OtpVerification:
    properties:
      otp:
        type: string
    required:
    - otp
    type: object
...
  requestmodels.RestaurantLoginReq:
    properties:
      password:
        minLength: 4
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
  requestmodels.RestaurantSignUpReq:
    properties:
      confirmPassword:
        type: string
      description:
        minLength: 10
        type: string
      district:
        minLength: 2
        type: string
      email:
        type: string
      gst_no:
        minLength: 2
        type: string
      locality:
        minLength: 2
        type: string
      owner_name:
        minLength: 2
        type: string
      password:
        minLength: 4
        type: string
      phone:
        type: string
      pinCode:
        type: string
      restaurant_name:
        minLength: 2
        type: string
    required:
    - confirmPassword
    - description
    - district
    - email
    - gst_no
    - locality
    - owner_name
    - password
    - phone
    - pinCode
    - restaurant_name
    type: object
...
  requestmodels.UserLoginReq:
    properties:
      password:
        minLength: 4
        type: string
      phone:
        minLength: 10
        type: string
    required:
    - password
    - phone
    type: object
  requestmodels.UserSignUpReq:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      firstName:
        minLength: 3
        type: string
      lastName:
        minLength: 1
        type: string
      password:
        minLength: 3
        type: string
      phone:
        type: string
    required:
    - confirmPassword
    - email
    - firstName
    - lastName
    - password
    - phone
    type: object
...
  responsemodels.Response:
    properties:
      after exicution: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
  description: Online Veg Food Delivery.
  title: Go + Gin Veg*n.
  version: "1.0"
paths:
...
  /login/:
    post:
      consumes:
      - application/json
      description: User can login using this handler.
      parameters:
      - description: Login Credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requestmodels.UserLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      summary: UserLogin
      tags:
      - User
...
  /profile:
    get:
      consumes:
      - application/json
      description: Retrieves the profile information for a specific user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: GetUserProfile
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Edits a user's profile.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: EditUserProfile
      tags:
      - User
...
  /restaurant/login:
    post:
      consumes:
      - application/json
      description: Logs in a restaurant.
      parameters:
      - description: Restaurant login credentials.
        in: body
        name: RestaurantLoginData
        required: true
        schema:
          $ref: '#/definitions/requestmodels.RestaurantLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      summary: RestaurantLogin
      tags:
      - Restaurant
...
  /restaurant/signup:
    post:
      consumes:
      - application/json
      description: Creates a new restaurant account.
      parameters:
      - description: Restaurant signup data.
        in: body
        name: restaurantSignUpData
        required: true
        schema:
          $ref: '#/definitions/requestmodels.RestaurantSignUpReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      summary: RestaurantSignUp
      tags:
      - Restaurant
  /signup/:
    post:
      consumes:
      - application/json
      description: User can signup using this handler
      parameters:
      - description: User Sign-Up Details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requestmodels.UserSignUpReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      summary: UserSignUp
      tags:
      - User
  /verify/:
    post:
      consumes:
      - application/json
      description: User can verify the OTP which is generated after successful signup
        request.
      parameters:
      - description: OTP generated
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requestmodels.OtpVerification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - OtpTempTokenAuth: []
      summary: UserOTPVerication
      tags:
      - User
securityDefinitions:
  AdminRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
  OtpTempTokenAuth:
    in: header
    name: Authorizations
    type: apiKey
  RestaurantAuthTokenAuth:
    in: header
    name: accesstoken
    type: apiKey
  RestaurantRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
  UserAuthTokenAuth:
    in: header
    name: accesstoken
    type: apiKey
  UserRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
swagger: "2.0"